Move all FixnumNodes in IntegerNodes
