+++ src/main/java/org/truffleruby/core/inlined/InlinedLeftShiftNode.java
@@ -10,8 +10,8 @@
 package org.truffleruby.core.inlined;
 
 import org.truffleruby.RubyContext;
-import org.truffleruby.core.numeric.FixnumNodes.LeftShiftNode;
-import org.truffleruby.core.numeric.FixnumNodesFactory.LeftShiftNodeFactory;
+import org.truffleruby.core.numeric.IntegerNodes.LeftShiftNode;
+import org.truffleruby.core.numeric.IntegerNodesFactory.LeftShiftNodeFactory;
 import org.truffleruby.language.dispatch.RubyCallNodeParameters;
 
 import com.oracle.truffle.api.CompilerDirectives;
