+++ src/main/java/org/truffleruby/core/inlined/InlinedSubNode.java
@@ -10,8 +10,8 @@
 package org.truffleruby.core.inlined;
 
 import org.truffleruby.RubyContext;
-import org.truffleruby.core.numeric.FixnumNodes.SubNode;
-import org.truffleruby.core.numeric.FixnumNodesFactory.SubNodeFactory;
+import org.truffleruby.core.numeric.IntegerNodes.SubNode;
+import org.truffleruby.core.numeric.IntegerNodesFactory.SubNodeFactory;
 import org.truffleruby.language.dispatch.RubyCallNodeParameters;
 
 import com.oracle.truffle.api.CompilerDirectives;
